+++ crates/zng-view/src/config/macos.rs
@@ -1,7 +1,9 @@
 use objc2_app_kit::*;
 use objc2_foundation::*;
 use zng_unit::{Rgba, TimeUnits as _};
-use zng_view_api::config::{AnimationsConfig, ColorScheme, ColorsConfig, FontAntiAliasing, KeyRepeatConfig, MultiClickConfig, TouchConfig};
+use zng_view_api::config::{
+    AnimationsConfig, ChromeConfig, ColorScheme, ColorsConfig, FontAntiAliasing, KeyRepeatConfig, MultiClickConfig, TouchConfig,
+};
 
 pub fn font_aa() -> FontAntiAliasing {
     super::other::font_aa()
@@ -71,6 +73,10 @@ pub fn colors_config() -> ColorsConfig {
 
 pub fn locale_config() -> zng_view_api::config::LocaleConfig {
     super::other::locale_config()
+}
+
+pub fn chrome_config() -> ChromeConfig {
+    ChromeConfig::default()
 }
 
 pub fn spawn_listener(l: crate::AppEventSender) -> Option<Box<dyn FnOnce()>> {