+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+* Add better custom chrome for GNOME+Wayland.
+* Add `zng::widget::node::bind_state_init` helper.
+* Add `Window::needs_fallback_chrome` and `Window::prefer_custom_chrome` property.
+* **Breaking** Add `ChromeConfig` and related events to the view API.
+    - Note that this is only a breaking change for direct dependents of `zng-view-api`.
 
 # 0.11.6